requirements.txt
torch
torchvision
PyYAML
matplotlib
Pillow
opencv-python
